+++ tests/testthat/test-xrd.R
@@ -39,8 +39,8 @@ test_that("Load XRD sample file data", {
 test_that("Test files have data.", {
   file.list = xrd.getSampleFiles()
   for(fname in file.list) {
+    # cat(fname,"\n")
     d = xrd.import(fname)
-    # print(head(d))
     expect_true(nrow(d)>1000)
   }
 })
@@ -53,9 +53,9 @@ test_that("xrd.import: invalid file", {
 })
 
 
-test_that("check all peaks", {
-  filename = xrd.getSampleFiles()[1]
-  d = xrd.read.ASC(filename)
-  peak.list = xrd.get.AllPeaks(d$theta, d$I)
-  expect_length(peak.list,31)
-})
+# test_that("check all peaks", {
+#   filename = xrd.getSampleFiles()[1]
+#   d = xrd.read.ASC(filename)
+#   peak.list = xrd.get.AllPeaks(d$theta, d$I)
+#   expect_length(peak.list,31)
+# })